feat(core): improved settings developer sections ui
