--- frontend/core-ui/src/modules/logs/components/filters/LogRecordTableFilterDialogs.tsx
@@ -1,11 +1,0 @@
-import { Filter } from 'erxes-ui';
-
-export const LogRecordTableFilterDialogs = () => {
-  return (
-    <Filter.Dialog>
-      <Filter.View filterKey="createdAt" inDialog>
-        <Filter.DialogDateView filterKey="createdAt" />
-      </Filter.View>
-    </Filter.Dialog>
-  );
-};
