+++ frontend/core-ui/src/modules/logs/components/filters/LogStatusFilter.tsx
@@ -1,9 +1,20 @@
-import {
-  IconCheck,
-  IconProgressCheck,
-  IconProgressX,
-} from '@tabler/icons-react';
-import { Command, Combobox, useMultiQueryState } from 'erxes-ui';
+import { IconCheck, IconProgressCheck, IconProgressX } from '@tabler/icons-react';
+import { Combobox, Command, useMultiQueryState } from 'erxes-ui';
+
+import { ILogStatusType } from '@/logs/types';
+
+const LOG_STATUS_OPTIONS = [
+  {
+    value: ILogStatusType.SUCCESS,
+    label: 'Success',
+    icon: IconProgressCheck,
+  },
+  {
+    value: ILogStatusType.FAILED,
+    label: 'Failed',
+    icon: IconProgressX,
+  },
+] as const;
 
 export const LogStatusFilter = () => {
   const [queries, setQueries] = useMultiQueryState<{
@@ -14,28 +25,13 @@ export const LogStatusFilter = () => {
 
   return (
     <Command shouldFilter={false}>
-      <Command.List className="p-1 ">
+      <Command.List className="p-1">
         <Combobox.Empty />
-        {[
-          {
-            value: 'success',
-            className:
-              'bg-success/10 data-[selected=true]:bg-success/20 text-success border-success/10',
-            icon: IconProgressCheck,
-            label: 'Success',
-          },
-          {
-            value: 'failed',
-            className:
-              'mt-2 bg-destructive/10 data-[selected=true]:bg-destructive/20 text-destructive border-destructive/10',
-            icon: IconProgressX,
-            label: 'Failed',
-          },
-        ].map(({ value, className, icon: Icon, label }) => (
+        {LOG_STATUS_OPTIONS.map(({ value, label, icon: Icon }) => (
           <Command.Item
             key={value}
             value={value}
-            className={`cursor-pointer ${className}`}
+            className="cursor-pointer"
             onSelect={() =>
               setQueries({
                 status: value === status ? null : value,
