--- frontend/core-ui/src/modules/logs/components/filters/LogUsersFilter.tsx
@@ -1,5 +1,0 @@
-import { SelectMember } from 'ui-modules';
-
-export const LogUsersFilter = () => {
-  return <SelectMember.FilterView queryKey="userIds" />;
-};
