+++ frontend/core-ui/src/modules/logs/components/filters/LogsRecordTableFilter.tsx
@@ -1,44 +1,16 @@
-import { Combobox, Filter, useMultiQueryState } from 'erxes-ui';
-import { LogRecordTableFilterBars } from './LogRecordTableFilterBars';
-import { LogRecordTableFilterDialogs } from './LogRecordTableFilterDialogs';
-import { LogRecordTableFilterMenu } from './LogRecordTableFilterMenu';
-import { LogRecordTableFilterViews } from './LogRecordTableFilterViews';
+import { Filter } from 'erxes-ui';
+
 import { LOGS_CURSOR_SESSION_KEY } from '@/logs/constants/logFilter';
+import { LogsFilterBar } from './LogsFilterBar';
+import { LogsFilterPopover } from './LogsFilterPopover';
 
 export const LogsRecordTableFilter = () => {
-  const [queries] = useMultiQueryState<{
-    status: string;
-    source: string;
-    action: string;
-    userIds: string[];
-    createdAt: string;
-    contentType: string;
-    docId: string;
-  }>([
-    'status',
-    'source',
-    'action',
-    'userIds',
-    'createdAt',
-    'contentType',
-    'docId',
-  ]);
-
-  const hasFilters = Object.values(queries || {}).some(
-    (value) => value !== null,
-  );
-
   return (
-    <Filter id="customers-filter" sessionKey={LOGS_CURSOR_SESSION_KEY}>
-      <Filter.Popover scope={'logs_page'}>
-        <Filter.Trigger isFiltered={hasFilters} />
-        <Combobox.Content>
-          <LogRecordTableFilterMenu />
-          <LogRecordTableFilterViews />
-        </Combobox.Content>
-      </Filter.Popover>
-      <LogRecordTableFilterDialogs />
-      <LogRecordTableFilterBars />
+    <Filter id="logs-filter" sessionKey={LOGS_CURSOR_SESSION_KEY}>
+      <Filter.Bar>
+        <LogsFilterPopover />
+        <LogsFilterBar />
+      </Filter.Bar>
     </Filter>
   );
 };
