+++ frontend/core-ui/src/modules/settings/oauth-clients/components/table/OAuthClientsSettingsColumns.tsx
@@ -5,37 +5,36 @@ import {
 } from '@/settings/oauth-clients/types';
 import { ColumnDef } from '@tanstack/table-core';
 import { format } from 'date-fns';
-import { Badge, RecordTableInlineCell, useToast } from 'erxes-ui';
-import { IconCopy } from '@tabler/icons-react';
+import {
+  Badge,
+  Button,
+  RecordTable,
+  RecordTableInlineCell,
+  useToast,
+} from 'erxes-ui';
+import { IconCopy, IconCheck } from '@tabler/icons-react';
 import { useState } from 'react';
 
 const ClientIdCell = ({ clientId }: { clientId: string }) => {
-  const [hovered, setHovered] = useState(false);
+  const [copied, setCopied] = useState(false);
   const { toast } = useToast();
 
   const handleCopy = () => {
+    setCopied(true);
     navigator.clipboard.writeText(clientId);
     toast({ variant: 'success', title: 'Client ID copied to clipboard' });
+    setTimeout(() => setCopied(false), 1000);
   };
 
   return (
-    <RecordTableInlineCell>
-      <div
-        className="flex items-center gap-1 group"
-        onMouseEnter={() => setHovered(true)}
-        onMouseLeave={() => setHovered(false)}
-      >
-        <span className="font-mono text-xs">{clientId}</span>
-        {hovered && (
-          <button
-            onClick={handleCopy}
-            className="p-0.5 rounded hover:bg-muted text-muted-foreground hover:text-foreground transition-colors"
-            title="Copy client ID"
-          >
-            <IconCopy size={14} />
-          </button>
-        )}
-      </div>
+    <RecordTableInlineCell
+      className="group justify-between"
+      onClick={handleCopy}
+    >
+      <span className="font-mono text-xs">{clientId}</span>
+      <Button size="icon" variant="ghost" className="group-hover:block hidden">
+        {copied ? <IconCheck size={14} /> : <IconCopy size={14} />}
+      </Button>
     </RecordTableInlineCell>
   );
 };
@@ -49,6 +48,7 @@ const getAccessTokenLifetimeLabel = (
 };
 
 export const oauthClientsSettingsColumns: ColumnDef<IOAuthClientApp>[] = [
+  { ...RecordTable.checkboxColumn, size: 20 } as ColumnDef<IOAuthClientApp>,
   {
     id: 'name',
     accessorKey: 'name',
