+++ frontend/core-ui/src/pages/settings/logs/LogsIndexPage.tsx
@@ -1,9 +1,15 @@
+import { LogsBreadcrumb } from '@/logs/components/LogsBreadcrumb';
 import { LogsRecordTable } from '@/logs/components/LogsRecordTable';
-import { PageContainer } from 'erxes-ui';
+import { LogsRecordTableFilter } from '@/logs/components/filters/LogsRecordTableFilter';
+import { PageContainer, PageSubHeader } from 'erxes-ui';
 
 export const LogsIndexPage = () => {
   return (
     <PageContainer>
+      <LogsBreadcrumb />
+      <PageSubHeader>
+        <LogsRecordTableFilter />
+      </PageSubHeader>
       <LogsRecordTable />
     </PageContainer>
   );
